rewrite tests to use shared header.
t
